+++ src/main/java/org/sonar/plugins/delphi/antlr/ast/node/FieldSectionNode.java
@@ -19,6 +19,7 @@
 package org.sonar.plugins.delphi.antlr.ast.node;
 
 import org.antlr.runtime.Token;
+import org.sonar.plugins.delphi.antlr.DelphiLexer;
 import org.sonar.plugins.delphi.antlr.ast.visitors.DelphiParserVisitor;
 
 public final class FieldSectionNode extends DelphiNode implements Visibility {
@@ -38,5 +39,9 @@ public final class FieldSectionNode extends DelphiNode implements Visibility {
   @Override
   public VisibilityType getVisibility() {
     return ((VisibilitySectionNode) jjtGetParent()).getVisibility();
+  }
+
+  public boolean isClassFieldSection() {
+    return getFirstChildWithId(DelphiLexer.CLASS) != null;
   }
 }