+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Exclude record class variables in `VariableInitializationRule`.
+
 ## [0.39.0] - 2022-11-11
 
 ### Added